+++ Sources/StreamChatSwiftUI/ChatChannel/ChatChannelView.swift
@@ -22,7 +22,7 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
     private var factory: Factory
 
     public init(
-        viewFactory: Factory,
+        viewFactory: Factory = DefaultViewFactory.shared,
         viewModel: ChatChannelViewModel? = nil,
         channelController: ChatChannelController,
         messageController: ChatMessageController? = nil,
@@ -86,6 +86,9 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
                         .if(viewModel.reactionsShown, transform: { view in
                             view.navigationBarHidden(true)
                         })
+                        .if(!viewModel.reactionsShown, transform: { view in
+                            view.navigationBarHidden(false)
+                        })
                         .if(viewModel.channelHeaderType == .regular) { view in
                             view.modifier(factory.makeChannelHeaderViewModifier(for: channel))
                         }
@@ -95,6 +98,7 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
                         .if(viewModel.channelHeaderType == .messageThread) { view in
                             view.modifier(factory.makeMessageThreadHeaderViewModifier())
                         }
+                        .animation(nil)
 
                     factory.makeMessageComposerViewType(
                         with: viewModel.channelController,
@@ -105,7 +109,7 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
                     )
                     .opacity((
                         utils.messageListConfig.messagePopoverEnabled && messageDisplayInfo != nil && !viewModel
-                            .reactionsShown
+                            .reactionsShown && viewModel.channel?.isFrozen == false
                     ) ? 0 : 1)
 
                     NavigationLink(