+++ Sources/StreamChatSwiftUI/ChatChannel/ChatChannelViewModel.swift
@@ -3,7 +3,6 @@
 //
 
 import Combine
-import Nuke
 import StreamChat
 import SwiftUI
 
@@ -37,6 +36,8 @@ open class ChatChannelViewModel: ObservableObject, MessagesDataSource {
     private var lastMessageRead: String?
     private var disableDateIndicator = false
     private var channelName = ""
+    private var onlineIndicatorShown = false
+    private let throttler = Throttler(interval: 3, broadcastLatestEvent: true)
     
     public var channelController: ChatChannelController
     public var messageController: ChatMessageController?
@@ -59,7 +60,9 @@ open class ChatChannelViewModel: ObservableObject, MessagesDataSource {
     @Published public var currentSnapshot: UIImage? {
         didSet {
             withAnimation {
-                reactionsShown = currentSnapshot != nil && utils.messageListConfig.messagePopoverEnabled
+                reactionsShown = currentSnapshot != nil
+                    && utils.messageListConfig.messagePopoverEnabled
+                    && channel?.isFrozen == false
             }
         }
     }
@@ -88,6 +91,7 @@ open class ChatChannelViewModel: ObservableObject, MessagesDataSource {
             if threadMessageShown == false {
                 threadMessage = nil
             }
+            utils.messageCachingUtils.messageThreadShown = threadMessageShown
         }
     }
 
@@ -158,7 +162,7 @@ open class ChatChannelViewModel: ObservableObject, MessagesDataSource {
     
     @objc
     private func didReceiveMemoryWarning() {
-        Nuke.ImageCache.shared.removeAll()
+        ImageCache.shared.removeAll()
         messageCachingUtils.clearCache()
     }
     
@@ -271,6 +275,7 @@ open class ChatChannelViewModel: ObservableObject, MessagesDataSource {
         checkReadIndicators(for: channel)
         checkTypingIndicator()
         checkHeaderType()
+        checkOnlineIndicator()
     }
 
     public func showReactionOverlay(for view: AnyView) {
@@ -338,7 +343,9 @@ open class ChatChannelViewModel: ObservableObject, MessagesDataSource {
     private func maybeSendReadEvent(for message: ChatMessage) {
         if message.id != lastMessageRead {
             lastMessageRead = message.id
-            channelController.markRead()
+            throttler.throttle { [weak self] in
+                self?.channelController.markRead()
+            }
         }
     }
     
@@ -377,13 +384,29 @@ open class ChatChannelViewModel: ObservableObject, MessagesDataSource {
         }
         
         if nameChanged {
-            // Toolbar is not updated unless there's a state change.
-            // Therefore, we manually need to update the state for a short period of time.
-            let headerType = channelHeaderType
-            channelHeaderType = .typingIndicator
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) { [weak self] in
-                self?.channelHeaderType = headerType
-            }
+            triggerHeaderChange()
+        }
+    }
+    
+    private func checkOnlineIndicator() {
+        guard let channel else { return }
+        let updated = !channel.lastActiveMembers.filter { member in
+            member.id != chatClient.currentUserId && member.isOnline
+        }.isEmpty
+        
+        if updated != onlineIndicatorShown {
+            onlineIndicatorShown = updated
+            triggerHeaderChange()
+        }
+    }
+    
+    private func triggerHeaderChange() {
+        // Toolbar is not updated unless there's a state change.
+        // Therefore, we manually need to update the state for a short period of time.
+        let headerType = channelHeaderType
+        channelHeaderType = .typingIndicator
+        DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) { [weak self] in
+            self?.channelHeaderType = headerType
         }
     }
     
@@ -487,7 +510,7 @@ open class ChatChannelViewModel: ObservableObject, MessagesDataSource {
         messageCachingUtils.clearCache()
         if messageController == nil {
             utils.channelControllerFactory.clearCurrentController()
-            Nuke.ImageCache.shared.trim(toCost: utils.messageListConfig.cacheSizeOnChatDismiss)
+            ImageCache.shared.trim(toCost: utils.messageListConfig.cacheSizeOnChatDismiss)
         }
     }
 }