+++ Sources/StreamChatSwiftUI/ChatChannel/Reactions/ReactionsOverlayView.swift
@@ -14,10 +14,15 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
     @State private var popIn = false
     @State private var willPopOut = false
     @State private var screenHeight = UIScreen.main.bounds.size.height
+    @State private var screenWidth: CGFloat?
+    @State private var initialWidth: CGFloat?
+    @State private var orientationChanged = false
+    @State private var initialOrigin: CGFloat?
 
     var factory: Factory
     var channel: ChatChannel
     var currentSnapshot: UIImage
+    var bottomOffset: CGFloat
     var messageDisplayInfo: MessageDisplayInfo
     var onBackgroundTap: () -> Void
     var onActionExecuted: (MessageActionInfo) -> Void
@@ -34,6 +39,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
         channel: ChatChannel,
         currentSnapshot: UIImage,
         messageDisplayInfo: MessageDisplayInfo,
+        bottomOffset: CGFloat = 0,
         onBackgroundTap: @escaping () -> Void,
         onActionExecuted: @escaping (MessageActionInfo) -> Void
     ) {
@@ -45,6 +51,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
         self.channel = channel
         self.factory = factory
         self.currentSnapshot = currentSnapshot
+        self.bottomOffset = bottomOffset
         self.messageDisplayInfo = messageDisplayInfo
         self.onBackgroundTap = onBackgroundTap
         self.onActionExecuted = onActionExecuted
@@ -58,11 +65,17 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
 
     public var body: some View {
         ZStack(alignment: .topLeading) {
-            factory.makeReactionsBackgroundView(
-                currentSnapshot: currentSnapshot,
-                popInAnimationInProgress: !popIn
-            )
-            .offset(y: spacing > 0 ? screenHeight - currentSnapshot.size.height : 0)
+            ZStack {
+                if !orientationChanged {
+                    factory.makeReactionsBackgroundView(
+                        currentSnapshot: currentSnapshot,
+                        popInAnimationInProgress: !popIn
+                    )
+                    .offset(y: spacing > 0 ? screenHeight - currentSnapshot.size.height : 0)
+                } else {
+                    Color.gray.opacity(0.4)
+                }
+            }
             .transition(.opacity)
             .onTapGesture {
                 dismissReactionsOverlay() { /* No additional handling. */ }
@@ -72,7 +85,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
                 Alert.defaultErrorAlert
             }
 
-            if !messageDisplayInfo.message.isSentByCurrentUser &&
+            if !messageDisplayInfo.message.isRightAligned &&
                 utils.messageListConfig.messageDisplayOptions.showAvatars(for: channel) {
                 factory.makeMessageAvatarView(
                     for: utils.messageCachingUtils.authorInfo(from: messageDisplayInfo.message)
@@ -87,7 +100,10 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
             GeometryReader { reader in
                 let frame = reader.frame(in: .local)
                 let height = frame.height
+                let width = frame.width
                 Color.clear.preference(key: HeightPreferenceKey.self, value: height)
+                Color.clear.preference(key: WidthPreferenceKey.self, value: width)
+                
                 VStack(alignment: .leading) {
                     Group {
                         if messageDisplayInfo.frame.height > messageContainerHeight {
@@ -113,7 +129,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
                     .scaleEffect(popIn || willPopOut ? 1 : 0.95)
                     .animation(willPopOut ? .easeInOut : popInAnimation, value: popIn)
                     .offset(
-                        x: messageDisplayInfo.frame.origin.x - diffWidth(proxy: reader)
+                        x: messageOriginX(proxy: reader)
                     )
                     .overlay(
                         channel.config.reactionsEnabled ?
@@ -130,7 +146,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
                             .opacity(willPopOut ? 0 : 1)
                             .animation(willPopOut ? .easeInOut : popInAnimation, value: popIn)
                             .offset(
-                                x: messageDisplayInfo.frame.origin.x - diffWidth(proxy: reader),
+                                x: messageOriginX(proxy: reader),
                                 y: popIn ? -24 : -messageContainerHeight / 2
                             )
                             .accessibilityElement(children: .contain)
@@ -162,7 +178,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
                         .opacity(willPopOut ? 0 : 1)
                         .scaleEffect(popIn ? 1 : (willPopOut ? 0.4 : 0))
                         .animation(willPopOut ? .easeInOut : popInAnimation, value: popIn)
-                    } else {
+                    } else if messageDisplayInfo.showsBottomContainer {
                         factory.makeReactionsUsersView(
                             message: viewModel.message,
                             maxHeight: userReactionsHeight
@@ -179,6 +195,9 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
                     }
                 }
                 .offset(y: !popIn ? (messageDisplayInfo.frame.origin.y - spacing) : originY)
+                .onAppear {
+                    self.initialOrigin = messageDisplayInfo.frame.origin.x - diffWidth(proxy: reader)
+                }
             }
         }
         .onPreferenceChange(HeightPreferenceKey.self) { value in
@@ -186,13 +205,24 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
                 self.screenHeight = value
             }
         }
+        .onPreferenceChange(WidthPreferenceKey.self) { value in
+            if initialWidth == nil {
+                initialWidth = value
+            }
+            self.screenWidth = value
+        }
         .edgesIgnoringSafeArea(.all)
-        .background(Color(colors.background))
+        .background(orientationChanged ? nil : Color(colors.background))
         .onAppear {
             popIn = true
         }
         .accessibilityElement(children: .contain)
         .accessibilityIdentifier("ReactionsOverlayView")
+        .onRotate { _ in
+            if isIPad {
+                self.orientationChanged = true
+            }
+        }
     }
 
     private func dismissReactionsOverlay(completion: @escaping () -> Void) {
@@ -208,14 +238,23 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
 
     private func messageActionsOffsetX(reader: GeometryProxy) -> CGFloat {
         let originX = messageActionsOriginX(availableWidth: reader.size.width)
-        let sentByCurrentUser = messageDisplayInfo.message.isSentByCurrentUser
         if popIn {
             return originX
         } else if willPopOut {
-            return messageDisplayInfo.frame.origin.x - diffWidth(proxy: reader)
+            return messageOriginX(proxy: reader)
         } else {
-            return sentByCurrentUser ? messageActionsWidth : 0
+            return messageDisplayInfo.message.isRightAligned ? messageActionsWidth : 0
         }
+    }
+        
+    private func messageOriginX(proxy: GeometryProxy) -> CGFloat {
+        let origin = messageDisplayInfo.frame.origin.x - diffWidth(proxy: proxy)
+        if let initialWidth, let initialOrigin, let screenWidth, abs(initialWidth - screenWidth) > 5 {
+            let diff = initialWidth - initialOrigin
+            let newOrigin = screenWidth - diff
+            return newOrigin
+        }
+        return initialOrigin ?? origin
     }
 
     private var messageContainerHeight: CGFloat {
@@ -242,7 +281,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
             messageDisplayInfo.showsMessageActions ? messageActionsSize : userReactionsPopupHeight
         var originY = messageDisplayInfo.frame.origin.y
         let minOrigin: CGFloat = 100
-        let maxOrigin: CGFloat = screenHeight - messageContainerHeight - bottomPopupOffset - minOrigin
+        let maxOrigin: CGFloat = screenHeight - messageContainerHeight - bottomPopupOffset - minOrigin - bottomOffset
         if originY < minOrigin {
             originY = minOrigin
         } else if originY > maxOrigin {
@@ -283,7 +322,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
     }
 
     private func messageActionsOriginX(availableWidth: CGFloat) -> CGFloat {
-        if messageDisplayInfo.message.isSentByCurrentUser {
+        if messageDisplayInfo.message.isRightAligned {
             return availableWidth - messageActionsWidth - paddingValue / 2
         } else {
             return CGSize.messageAvatarSize.width + paddingValue
@@ -291,7 +330,7 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
     }
 
     private func userReactionsOriginX(availableWidth: CGFloat) -> CGFloat {
-        if messageDisplayInfo.message.isSentByCurrentUser {
+        if messageDisplayInfo.message.isRightAligned {
             return availableWidth - maxUserReactionsWidth(availableWidth: availableWidth) - paddingValue / 2
         } else {
             return paddingValue
@@ -300,10 +339,28 @@ public struct ReactionsOverlayView<Factory: ViewFactory>: View {
 
     private var messageActionsWidth: CGFloat {
         var width = messageDisplayInfo.contentWidth + 2 * paddingValue
-        if messageDisplayInfo.message.isSentByCurrentUser {
+        if messageDisplayInfo.message.isRightAligned {
             width -= 2 * paddingValue
         }
 
         return width
+    }
+}
+
+struct DeviceRotationViewModifier: ViewModifier {
+    let action: (UIDeviceOrientation) -> Void
+
+    func body(content: Content) -> some View {
+        content
+            .onAppear()
+            .onReceive(NotificationCenter.default.publisher(for: UIDevice.orientationDidChangeNotification)) { _ in
+                action(UIDevice.current.orientation)
+            }
+    }
+}
+
+extension View {
+    func onRotate(perform action: @escaping (UIDeviceOrientation) -> Void) -> some View {
+        modifier(DeviceRotationViewModifier(action: action))
     }
 }